add a few primivite parsers + minor refactoring

+++ libs/common/inc/CppSip/message/sip_version.h
@@ -5,8 +5,8 @@
 namespace CppSip {
 
 struct SipVersion {
-  size_t major;
-  size_t minor;
+  unsigned int major;
+  unsigned int minor;
 };
 
 } // namespace CppSip
